Make it possible to override LogFilePath impl.
This is necessary for making proper DelimitedTextFile{Reader,Writer}
implementations as they need to know which offsets are in which file
(as the file path is determined by the configured MessageParser).

The current implementations will not handle trimming of log files
correctly, but for the sake of having the integration tests passing in
this repo I’ve left the functionality as-is, so it’s up to each user
if they want to implement a working solution.