Add key-based deduplication
A very simple and best-effort (heavy on the "effort") deduplication
implementation that just maintains the latest X number of keys per
partition and discards duplicated messages.

After a rebalance event the deduplicator will have to be reseted or
otherwise it'll incorrectly discard replayed messages (as a rebalance
event leads to a replay of the last X minutes of messages).